+++ src/simmate/toolkit/dataframes/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .molecule import MoleculeDataFrame