+++ src/simmate/toolkit/featurizers/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from .method_caller import MethodCaller
 from .morgan_fp import MorganFingerprint
+from .pattern_fingerprint import PatternFingerprint
 from .property_grabber import PropertyGrabber