+++ src/simmate/toolkit/featurizers/property_grabber.py
@@ -11,6 +11,12 @@ class PropertyGrabber(Featurizer):
     Grabs attributes from the molecule object
     """
 
+    @classmethod
+    def get_feature_names(cls, properties: list[str], **kwargs) -> list[str]:
+        # the list of property names is given by the user and we just copy it
+        # over so that higher level features can use it
+        return properties
+
     @staticmethod
     def featurize(
         molecule: Molecule,