+++ docs/change_log.md
@@ -77,6 +77,8 @@ There is one significant exception to the above rules -- `MAJOR=0` releases. Any
 - added dev apps `chatbot` and `analysis_dashboard`
 - added `database.external_connectors` utils
 - added proof of concept htmx app for frontend forms
+- added `toolkit.datastores` and `toolkit.dataframes` helper modules
+
 
 **Refactors**
 
@@ -86,6 +88,7 @@ There is one significant exception to the above rules -- `MAJOR=0` releases. Any
 - workflows results in the web ui now exists within the Data tab. The workflow tab has been limited to submission, docs, and analytics
 - `DynamicApiView` was split up into separate views for easier debugging and future enhancements
 - `website.engine` apps and others are condensed into `core_components`, `data_explorer`, and `workflows` apps
+- refactored `toolkit.featurizers` module
 
 **Fixes**